+++ client/src/components/pages/SearchPage.js
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
 import PageHeader from '../organisms/PageHeader';
-import FilterContainer from '../organisms/listcontents/ListFilter';
-import PaginationContainer from '../organisms/listcontents/ListPagination';
-import SeriesListContainer from '../organisms/listcontents/SeriesListContainer';
+// import FilterContainer from '../organisms/listContainter/ListFilter';
+// import PaginationContainer from '../organisms/listContainter/ListPagination';
+import SeriesListContainer from '../organisms/listContainter/SeriesListContainer';
 import SearchResultTitle from '../organisms/search/SearchResultTitle';
-import PostListContainer from '../organisms/listcontents/PostListContainer';
+import PostListContainer from '../organisms/listContainter/PostListContainer';
 
 const Container = styled.div`
   display: flex;
@@ -19,12 +19,12 @@ const Search = () => {
     <Container>
       <PageHeader />
       <SearchResultTitle title="검색결과" amount="120" />
-      <FilterContainer />
+      {/* <FilterContainer /> */}
       <SearchResultTitle title="Post" amount="60" />
       <PostListContainer />
       <SearchResultTitle title="Series" amount="60" />
       <SeriesListContainer />
-      <PaginationContainer />
+      {/* <PaginationContainer /> */}
     </Container>
   );
 };
