+++ client/src/components/pages/SeriesListPage.js
@@ -1,29 +1,30 @@
 import styled from 'styled-components';
 import PageHeader from '../organisms/PageHeader';
-import Pagination from '../molecules/Pagination';
+import SeriesListContainer from '../organisms/listContainter/SeriesListContainer';
 import Lnb from '../organisms/Lnb';
-import PostListContainer from '../organisms/listcontents/PostListContainer';
+import Pagination from '../molecules/Pagination';
 import SearchResultTitle from '../organisms/search/SearchResultTitle';
 
 const Container = styled.div`
   display: flex;
   flex-direction: column;
-  width: 1060px;
+  width: 100%;
   margin: 0 auto;
   justify-content: center;
   align-items: center;
 `;
 
-const PostList = () => {
+const SeriesList = () => {
   return (
     <Container>
       <PageHeader />
       <SearchResultTitle title="검색결과" amount="120" />
+
       <Lnb />
-      <PostListContainer />
-      <Pagination totalPages={30} />
+      <SeriesListContainer />
+      <Pagination totalPages={10} />
     </Container>
   );
 };
 
-export default PostList;
+export default SeriesList;
