+++ client/src/components/pages/SeriesPage.js
